requirements.txt
telethon
PyYAML
feedparser
fastapi
uvicorn
